Getting this merged so our branches don't diverge too much. I'd like to pull the Frankenboard class into its own file and maybe rename Controller.h/cpp and remove the Sample* stuff, but I'll do those in a separate commit that doesn't include logic changes so the merges aren't so confusing.

+++ hwhlk/TestPlugin/I2cDriver.h
@@ -18,10 +18,10 @@ class I2cDriver
 public:
     I2cDriver(winrt::Windows::Devices::Usb::UsbDevice usbDevice);
 
-    uint16_t readRegister(uint16_t address, uint8_t reg, size_t count, uint8_t* buffer);
+    uint16_t readRegister(uint16_t address, uint8_t reg, uint16_t count, uint8_t* buffer);
     char readRegisterByte(uint16_t address, uint8_t reg);
 
-    void writeRegister(uint16_t address, uint8_t reg, size_t count, const uint8_t* buffer);
+    void writeRegister(uint16_t address, uint8_t reg, uint32_t count, const uint8_t* buffer);
     void writeRegisterByte(uint16_t address, uint8_t reg, uint8_t value);
     void writeRegisterByteMasked(uint16_t address, uint8_t reg, uint8_t value, uint8_t mask);
